Moves selectMove from MiniMax to UnbeatableComputer class
